// frontend/src/types/api/fieldOperator.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * mailassist
 * OpenAPI spec version: 0.1.0
 */

/**
 * Comparison operators for individual condition rules.
 */
export type FieldOperator = typeof FieldOperator[keyof typeof FieldOperator];


export const FieldOperator = {
  equals: 'equals',
  not_equals: 'not_equals',
  contains: 'contains',
  not_contains: 'not_contains',
  starts_with: 'starts_with',
  ends_with: 'ends_with',
  matches_regex: 'matches_regex',
  greater_than: 'greater_than',
  less_than: 'less_than',
  is_empty: 'is_empty',
  is_not_empty: 'is_not_empty',
} as const;